+++ intern/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+
+intern = Blueprint('errors', __name__)
+
