+++ templates/bass.html
@@ -47,7 +47,10 @@
                 </li>  
               <li class="nav-item">
                 <a class="nav-link" href="/intern">Intern</a>
-              </li>    
+              </li>
+              <li class="nav-item">
+                <a class="nav-link" href="/intern/admin">Admin</a>
+              </li>     
             </ul>
           </div>  
         </nav>
