+++ templates/intern.html
@@ -36,7 +36,7 @@
                 {% endif %}
       
                   <th scope="row">{{ loop.index }}</th>
-                  <td><a href="/entry/{{ file[0] }}">{{ file[1] }}</a></td>
+                  <td><a href="/intern/entry/{{ file[0] }}">{{ file[1] }}</a></td>
                   <td >{{ file[3] }}</td>
                   <td>{{ file[4] }}</td>
